+++ src/CG.Options/GlobalUsings.cs
@@ -1,0 +1,14 @@
+﻿global using CG.DataAnnotations;
+global using CG.Options;
+global using CG.Validations;
+global using Microsoft.Extensions.Configuration;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.DependencyInjection.Extensions;
+global using Microsoft.Extensions.Options;
+global using System;
+global using System.Collections.Generic;
+global using System.ComponentModel.DataAnnotations;
+global using System.Linq;
+global using System.Runtime.CompilerServices;
+global using System.Runtime.Serialization;
+global using System.Reflection;